requirements.txt
accelerate==0.28.0
beautifulsoup4==4.12.3
bitsandbytes==0.43.0
blinker==1.4
cbor==1.0.0
certifi==2024.2.2
charset-normalizer==3.3.2
chest==0.2.3
command-not-found==0.3
cryptography==3.4.8
cwl-eval==1.0.12
dbus-python==1.2.18
Deprecated==1.2.14
dill==0.3.8
distro==1.7.0
distro-info==1.1+ubuntu0.1
filelock==3.13.3
fsspec==2024.3.1
HeapDict==1.0.1
httplib2==0.20.2
huggingface-hub==0.22.2
idna==3.6
ijson==3.2.3
importlib-metadata==4.6.4
inscriptis==2.5.0
ir-datasets==0.5.6
ir-measures==0.3.3
jeepney==0.7.1
Jinja2==3.1.3
joblib==1.3.2
keyring==23.5.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lxml==5.2.1
lz4==4.3.3
MarkupSafe==2.1.5
matchpy==0.5.5
more-itertools==8.10.0
mpmath==1.3.0
multiset==2.1.1
netifaces==0.11.0
networkx==3.2.1
nptyping==1.4.4
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.0
packaging==24.0
pandas==2.2.1
patsy==0.5.6
psutil==5.9.8
pyautocorpus==0.1.12
PyGObject==3.42.1
pyjnius==1.6.1
PyJWT==2.3.0
pyparsing==2.4.7
python-apt==2.4.0+ubuntu2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-terrier==0.10.0
pytrec-eval-terrier==0.5.6
pytz==2024.1
PyYAML==5.4.1
regex==2023.12.25
requests==2.31.0
safetensors==0.4.2
scikit-learn==1.4.1.post1
scipy==1.13.0
SecretStorage==3.3.1
sentencepiece==0.2.0
six==1.16.0
soupsieve==2.5
statsmodels==0.14.1
sympy==1.12
systemd-python==234
threadpoolctl==3.4.0
tokenizers==0.15.2
torch==2.2.2
tqdm==4.66.2
transformers==4.39.3
trec-car-tools==2.6
triton==2.2.0
typing_extensions==4.10.0
typish==1.9.3
tzdata==2024.1
ubuntu-advantage-tools==8001
ufw==0.36.1
unattended-upgrades==0.1
unlzw3==0.2.2
urllib3==2.2.1
wadllib==1.3.6
warc3-wet==0.2.3
warc3-wet-clueweb09==0.2.5
wget==3.2
wrapt==1.16.0
zipp==1.0.0
zlib-state==0.1.6
